extents: separate exceptions
The exception is required to be
caught by generate.py, so put
it in its own module.

Refs #1100
